refactor: create user/repositories dir & move UserRepository to this directory
