+++ src/modules/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
-import { UsersRepository } from './users.repository';
+import { UsersRepository } from './repositories/users.repository';
 
 @Module({
   exports: [UsersService],
